requirements.txt
# HireAI — Python Dependencies
# ✅ Requires Python 3.11 ONLY (3.12/3.13/3.14 NOT supported by PyTorch)
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
flask-pymongo==2.3.0
flask-caching==2.3.0
python-dotenv==1.0.1
werkzeug==3.0.3

# ── MongoDB Driver ─────────────────────────────────────────────────────────────
pymongo[srv]==4.7.3
dnspython==2.6.1

# ── ML Core ────────────────────────────────────────────────────────────────────
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3

# ── PyTorch (Python 3.11 compatible — installed via setup.sh/setup.bat) ────────
# DO NOT install torch here — run setup.sh or setup.bat which handles this
# torch==2.3.1  <-- installed separately with correct index URL

# ── NLP & Embeddings ───────────────────────────────────────────────────────────
sentence-transformers==2.7.0
transformers==4.40.2
tokenizers==0.19.1
huggingface-hub==0.23.4

# ── spaCy (NER) ────────────────────────────────────────────────────────────────
spacy==3.7.5
# After install run: python -m spacy download en_core_web_sm

# ── PDF Parsing ────────────────────────────────────────────────────────────────
PyMuPDF==1.24.5

# ── Explainability ─────────────────────────────────────────────────────────────
shap==0.45.1

# ── Utilities ──────────────────────────────────────────────────────────────────
requests==2.32.3
certifi==2024.7.4
tqdm==4.66.4
